task-04/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.8
lxml==5.3.0
pyTelegramBotAPI==4.22.1
python-docx==1.1.2
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.2
